Remove x509 template from API
